requirements.txt
Flask==2.2.2
python-telegram-bot==13.15
python-dotenv==0.21.0